feat: view venda operações crud e ajustes para otimização da aplicação

+++ backend_consorcio_automovel/venda/controllers.py
@@ -5,29 +5,39 @@ from automovel.models import Automovel
 
 class VendaController:
     @staticmethod
-    def criar_venda(data):
-        vendedor = User.objects.get(username=data.username_vendedor)
+    def criar_venda(request, data):
+        vendedor = User.objects.get(username=request.user.username)
         cliente = Cliente.objects.get(cpf=data.cpf_cliente)
         automovel = Automovel.objects.get(chassi=data.chassi_automovel)
+        preco = automovel.preco
+        
         venda = Venda.objects.create(
                 username_vendedor=vendedor,
                 cpf_cliente=cliente,
-                chassi_automovel=automovel
+                chassi_automovel=automovel,
+                preco=preco
         )
         return venda
     
     @staticmethod
-    def listar_vendas():
-        vendas = Venda.objects.all()
+    def listar_vendas(request):
+        if request.user.is_staff:
+            vendas = Venda.objects.all()
+        else:
+            vendas = Venda.objects.filter(username_vendedor=request.user)
         if not vendas:
             return None
         return list(vendas.values())
     
     @staticmethod
-    def deletar_venda(codigo_venda):
+    def deletar_venda(request, codigo_venda):
         try:
             venda = Venda.objects.get(codigo_venda=codigo_venda)
-            venda.delete()
+            if request.user.is_staff or venda.username_vendedor == request.user:
+                venda.delete()
+                return True
+            else:
+                return False
         except Venda.DoesNotExist:
             return None
